% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_generation.R
\name{create_one_package}
\alias{create_one_package}
\title{Create a package}
\usage{
create_one_package(df, items_per_package, materials)
}
\arguments{
\item{df}{dataframe containing the items we will sample from}

\item{items_per_package}{how many  items should the package contain}
}
\description{
Function to create one package
}
