% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut_stat_functions.R
\name{freedman_diaconis_binwidth}
\alias{freedman_diaconis_binwidth}
\title{Function to calculate bin width according to the
Freedman - Diaconis rule. Returns the number}
\usage{
freedman_diaconis_binwidth(x)
}
\arguments{
\item{x}{(vector) data to be plotted as histogram (eg. column of dataframe)}
}
\description{
Function to calculate bin width according to the
Freedman - Diaconis rule. Returns the number
}
