% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut_stat_functions.R
\name{freedman_diaconis_binwidth_log}
\alias{freedman_diaconis_binwidth_log}
\title{F-D binwidth logarithmic}
\usage{
freedman_diaconis_binwidth_log(x)
}
\arguments{
\item{x}{(vector) data to be plotted as histogram (eg. column of dataframe)}
}
\description{
Function to calculate bin width according to the
Freedman - Diaconis rule but in a log axis.
}
\details{
Returns the number.
}
