% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_geometry_wrappers.R
\name{geom_bar_rowiz}
\alias{geom_bar_rowiz}
\title{Wrapper function for geom_bar
Purpose: To include by default a smaller bar width because the default default one is too thick
Use: Exactly like geom_bar. Defaults will be overriden if specified otherwise in the arguments.}
\usage{
geom_bar_rowiz(
  mapping = NULL,
  data = NULL,
  ...,
  fill = "cadetblue",
  width = 0.4
)
}
\description{
Wrapper function for geom_bar
Purpose: To include by default a smaller bar width because the default default one is too thick
Use: Exactly like geom_bar. Defaults will be overriden if specified otherwise in the arguments.
}
