% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_geometry_wrappers.R
\name{geom_density_ridges_rowiz}
\alias{geom_density_ridges_rowiz}
\title{Wrapper function for geom_density_ridges
Purpose: To include by default a pre-defined linewidth, fill transparency and colour
Use: Exactly like geom_density_ridges. Defaults will be overriden if specified otherwise in the arguments.}
\usage{
geom_density_ridges_rowiz(
  mapping = NULL,
  data = NULL,
  ...,
  alpha = 0.6,
  linewidth = 0.8
)
}
\description{
Wrapper function for geom_density_ridges
Purpose: To include by default a pre-defined linewidth, fill transparency and colour
Use: Exactly like geom_density_ridges. Defaults will be overriden if specified otherwise in the arguments.
}
