% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_functions.R
\name{geom_histogram_custom}
\alias{geom_histogram_custom}
\title{Geometry to plot a histogram as a line
Use as a normal ggplot2 geometry by adding it
to your ggplot object. Use it like you would any geom_stuff,
so add it to ggplot objects.}
\usage{
geom_histogram_custom(variable, binwidth, color = "black")
}
\arguments{
\item{variable}{variable name to be plotted}

\item{binwidth}{width of the bins to be used}

\item{color}{colour of the line (defaults to black)}
}
\description{
Geometry to plot a histogram as a line
Use as a normal ggplot2 geometry by adding it
to your ggplot object. Use it like you would any geom_stuff,
so add it to ggplot objects.
}
