% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_geometry_wrappers.R
\docType{data}
\name{geom_lollipop_rowiz}
\alias{geom_lollipop_rowiz}
\alias{GeomLollipop}
\title{Custom ggplot style geometry to plot a lollipop plot
Purpose: To avoid manually adding points and segments etc
Use: Exactly like any other ggplot geom. Just use the x and y you want your points to be at
Defaults will be overriden if specified otherwise in the arguments.}
\format{
An object of class \code{GeomLollipop} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 4.
}
\usage{
GeomLollipop
}
\description{
Custom ggplot style geometry to plot a lollipop plot
Purpose: To avoid manually adding points and segments etc
Use: Exactly like any other ggplot geom. Just use the x and y you want your points to be at
Defaults will be overriden if specified otherwise in the arguments.
}
\keyword{datasets}
