% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_selectors.R
\name{get_radionuclides}
\alias{get_radionuclides}
\title{Extract the radionuclides present in the dataframe.
Returns a vector with the names.}
\usage{
get_radionuclides(df)
}
\arguments{
\item{df}{dataframe name}
}
\description{
Extract the radionuclides present in the dataframe.
Returns a vector with the names.
}
