% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_feature_engineering.r
\name{get_relevant_radionuclides}
\alias{get_relevant_radionuclides}
\title{Extract relevant radionuclides}
\usage{
get_relevant_radionuclides(df, limit_col)
}
\arguments{
\item{df}{dataframe name}

\item{limit_col}{the name of the column containing the limit we want to check (eg CL_lim)}
}
\description{
Find the relevant nuclides for a specific radiological limit.
By relevant nuclides we mean the ones that have an activity of >= 1\% of the whatever limit
for at least one item in the dataframe.
}
\details{
Returns a vector with the revelant nuclide names
}
