% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inspect_functions.R
\name{inspect_parameters}
\alias{inspect_parameters}
\title{Inspect all scenario parameters in a glance}
\usage{
inspect_parameters(df)
}
\arguments{
\item{df}{dataframe name}
}
\description{
Inspect all scenario parameters in a glance
}
