% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_functions.R
\name{label_machines}
\alias{label_machines}
\title{Label machines and group LHC experiments (keeps original column).
Returns the dataframe with an added column named machine_labelled.}
\usage{
label_machines(df)
}
\arguments{
\item{df}{dataframe name}
}
\description{
Label machines and group LHC experiments (keeps original column).
Returns the dataframe with an added column named machine_labelled.
}
