% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_input_parameters_single_plots.R
\name{plot_beam_losses}
\alias{plot_beam_losses}
\title{Function to plot the beam losses distribution
Returns the plot (ggplot)}
\usage{
plot_beam_losses(df)
}
\arguments{
\item{df}{dataframe with beam losses column (be it scn, itm or whatever works)}
}
\description{
Function to plot the beam losses distribution
Returns the plot (ggplot)
}
