% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_functions.R
\name{plot_char_quantities}
\alias{plot_char_quantities}
\title{Function to plot characterisation quantities: LL and IRAS.
Returns the plot (patchwork wrap).}
\usage{
plot_char_quantities(df)
}
\arguments{
\item{df}{dataframe with said quantities}
}
\description{
Function to plot characterisation quantities: LL and IRAS.
Returns the plot (patchwork wrap).
}
