% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_input_parameters_single_plots.R
\name{plot_dimensions}
\alias{plot_dimensions}
\title{Function to plot synthetic items' dimensions
Returns the plot (ggplot)}
\usage{
plot_dimensions(df)
}
\arguments{
\item{df}{dataframe with machine column (be it scn, itm or whatever works)}
}
\description{
Function to plot synthetic items' dimensions
Returns the plot (ggplot)
}
