% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_input_parameters_single_plots.R
\name{plot_filling_ratio}
\alias{plot_filling_ratio}
\title{Visualise filling ratio}
\usage{
plot_filling_ratio(df)
}
\arguments{
\item{df}{dataframe with filling ratio column (be it scn, itm or whatever works)}
}
\description{
Function to plot synthetic items' filing ratio as a density plot
}
\details{
Returns the plot (ggplot)
}
