% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_functions.R
\name{plot_irradiation}
\alias{plot_irradiation}
\title{Function to plot irradiation conditions: beam intensity,
irradiation time and waiting time. Returns the plot (patchwork wrap).}
\usage{
plot_irradiation(df)
}
\arguments{
\item{df}{Dataframe with beam intensity, irradiation and waiting time columns (can be scn or itm)}
}
\description{
Function to plot irradiation conditions: beam intensity,
irradiation time and waiting time. Returns the plot (patchwork wrap).
}
