% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_input_parameters_combined_plots.r
\name{plot_irradiation_parameters}
\alias{plot_irradiation_parameters}
\title{Visualise irradiation characteristics}
\usage{
plot_irradiation_parameters(df)
}
\arguments{
\item{df}{Dataframe with beam intensity, irradiation and waiting time columns (can be scn or itm)}
}
\description{
Function to plot irradiation characteristics: beam intensity,
irradiation time and waiting time, all as density plots.
}
\details{
Returns the plot (patchwork wrap).
}
