% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_input_parameters_single_plots.R
\name{plot_irradiation_time}
\alias{plot_irradiation_time}
\title{Function to plot the irradiation time distribution
Returns the plot (ggplot)}
\usage{
plot_irradiation_time(df)
}
\arguments{
\item{df}{dataframe with irradiation time column (be it scn, itm or whatever works)}
}
\description{
Function to plot the irradiation time distribution
Returns the plot (ggplot)
}
