% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_input_parameters_single_plots.R
\name{plot_machine}
\alias{plot_machine}
\title{Visualise machine of origin}
\usage{
plot_machine(df)
}
\arguments{
\item{df}{dataframe with machine column (be it scn, itm or whatever works)}
}
\description{
Function to plot machine of origin barlot with a percentage of each machine.
ATLAS, ALICE, CMS and LHCb are grouped together under LHC experiments.
}
\details{
Returns the plot (ggplot)
}
