% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_input_parameters_single_plots.R
\name{plot_mass}
\alias{plot_mass}
\title{Function to plot synthetic items' mass
Returns the plot (ggplot)}
\usage{
plot_mass(df)
}
\arguments{
\item{df}{dataframe with machine column (be it scn, itm or whatever works)}
}
\description{
Function to plot synthetic items' mass
Returns the plot (ggplot)
}
