% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_input_parameters_combined_plots.r
\name{plot_mass_info}
\alias{plot_mass_info}
\title{Visualise mass et al}
\usage{
plot_mass_info(df)
}
\arguments{
\item{df}{dataframe with mass, volume and density columns (be it scn or itm or whatever)}
}
\description{
Function to plot mass, volume and density info as density plots
}
\details{
Returns the plot (patchwork wrap)
}
