% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_input_parameters_single_plots.R
\name{plot_materials}
\alias{plot_materials}
\title{Function to plot material composition
Will plot main material distributions and subgroup percentages.
Returns the plot (patchwork wrap).}
\usage{
plot_materials(df)
}
\arguments{
\item{df}{dataframe containing a column with materials and one with groups (can be scn or itm). Note: column names are hard-coded but it will warn you if anything has changed.}
}
\description{
Function to plot material composition
Will plot main material distributions and subgroup percentages.
Returns the plot (patchwork wrap).
}
