% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_functions.R
\name{plot_measures}
\alias{plot_measures}
\title{Function to plot measurable quantities: bgo counting rate and
dose rate. Returns the plot (patchwork wrap)}
\usage{
plot_measures(df)
}
\arguments{
\item{df}{Dataframe with bgo and dose rate columns}
}
\description{
Function to plot measurable quantities: bgo counting rate and
dose rate. Returns the plot (patchwork wrap)
}
