% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_input_parameters_single_plots.R
\name{plot_position}
\alias{plot_position}
\title{Function to plot position of irradiation
Returns the plot (patchwork)}
\usage{
plot_position(df)
}
\arguments{
\item{df}{dataframe with position column (be it scn, itm or whatever works)}
}
\description{
Function to plot position of irradiation
Returns the plot (patchwork)
}
