% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_input_parameters_combined_plots.r
\name{plot_scenario_parameters}
\alias{plot_scenario_parameters}
\title{Visalise scenario parameters}
\usage{
plot_scenario_parameters(df)
}
\arguments{
\item{df}{Dataframe with scenario parameters (can be scn or itm, as long as it has columns with materials and irradiation details)}
}
\description{
Function to plot most importan scenario quantities:
Materials & irradiation details together.
You know, that plot that usually goes in a report.
}
\details{
Returns the plot (patchwork wrap)
}
