% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_input_parameters_single_plots.R
\name{plot_waiting_time}
\alias{plot_waiting_time}
\title{Visualise waiting time}
\usage{
plot_waiting_time(df)
}
\arguments{
\item{df}{dataframe with waiting time column (be it scn, itm or whatever works)}
}
\description{
Function to plot the waiting time distribution as a density plot per machine
}
\details{
Returns the plot (ggplot)
}
