% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_input_output.R
\name{read_input_dataframe}
\alias{read_input_dataframe}
\title{Accept an input csv file path and name and read it into a dataframe
making sure the options are what we need them to be:
Treat first row as header.
Don't convert strings to factors, keep them characters.
Don't "fix" the names .
Returns the dataframe}
\usage{
read_input_dataframe(filename)
}
\arguments{
\item{filename}{The full path and name of the input csv file}
}
\description{
Accept an input csv file path and name and read it into a dataframe
making sure the options are what we need them to be:
Treat first row as header.
Don't convert strings to factors, keep them characters.
Don't "fix" the names .
Returns the dataframe
}
