% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{save_plot}
\alias{save_plot}
\title{Exporting plots}
\usage{
save_plot(plot, rows = 1, columns = 1, legend = "n", filename)
}
\arguments{
\item{plot}{The plot we want to save}

\item{rows}{The number of plots in each column of the grid}

\item{columns}{The number of plots in each row of the grid}

\item{legend}{"y/n" Whether the plot includes a legend on the side (so it offsets the plot width to keep it looking nice)}

\item{filename}{The absolute path and filename of the plot}
}
\description{
Function that saves a plot, with its size depending on how many
subplots exist on the grid.
Aim: Have the same sizing for all
similar plots. Resolution is taken care of as well.
}
