% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_professional}
\alias{theme_professional}
\title{Theme so all plots are homogeneous in terms of
grids, axes, font sizes, etc. Use either by adding +theme_professional()
to your ggplot object or by theme_set(theme_professional()) in your script.}
\usage{
theme_professional(base_size = 12, base_family = "sans")
}
\description{
Theme so all plots are homogeneous in terms of
grids, axes, font sizes, etc. Use either by adding +theme_professional()
to your ggplot object or by theme_set(theme_professional()) in your script.
}
