% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_geometry_wrappers.R
\name{geom_histogram_rowiz}
\alias{geom_histogram_rowiz}
\title{Custom geom_histogram}
\usage{
geom_histogram_rowiz(
  mapping = NULL,
  data = NULL,
  ...,
  binwidth,
  color = "black",
  linewidth = 0.8
)
}
\description{
Purpose: To include by default a pre-defined linewidth and bin fill style
}
\details{
Use: Exactly like geom_histogram. Defaults will be overriden if specified otherwise in the arguments.
}
