% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_selectors.R
\name{get_materials}
\alias{get_materials}
\title{Extract material composition}
\usage{
get_materials(df)
}
\arguments{
\item{df}{dataframe name}
}
\description{
Check what are the main materials our items consist of
}
\details{
Returns a vector with the main material names
}
