% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_selectors.R
\name{get_radionuclides}
\alias{get_radionuclides}
\title{Extract radionuclides}
\usage{
get_radionuclides(df)
}
\arguments{
\item{df}{dataframe name}
}
\description{
Extract all the radionuclides present in the dataframe.
}
\details{
Returns a vector with the names.
}
