% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Inspect_functions.R
\name{inspect_parameters}
\alias{inspect_parameters}
\title{Inspect all scenario parameters in a glance
Returns a named list of plots: "plot_physical_characteristics"
and "plot_irradiation_characteristics"}
\usage{
inspect_parameters(df)
}
\arguments{
\item{df}{dataframe name}
}
\description{
Inspect all scenario parameters in a glance
Returns a named list of plots: "plot_physical_characteristics"
and "plot_irradiation_characteristics"
}
