% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_item_parameters.R
\name{plot_char_quantities}
\alias{plot_char_quantities}
\title{Visualise characterisation quantities}
\usage{
plot_char_quantities(df)
}
\arguments{
\item{df}{dataframe with said quantities}
}
\description{
Function to plot characterisation quantities: LL and IRAS.
}
\details{
Returns the plot (patchwork wrap).
}
