% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_input_parameters_single_plots.R
\name{plot_density}
\alias{plot_density}
\title{Function to plot synthetic items' density
Returns the plot (ggplot)}
\usage{
plot_density(df)
}
\arguments{
\item{df}{dataframe with density column (be it scn, itm or whatever works)}
}
\description{
Function to plot synthetic items' density
Returns the plot (ggplot)
}
