% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_characterised.R
\name{plot_IRAS}
\alias{plot_IRAS}
\alias{plot_iras}
\title{Visualise IRAS}
\usage{
plot_iras(df)
}
\arguments{
\item{df}{dataframe with characterised and aw iras}
}
\description{
Function to plot IRAS resulting from the characterisation
as well as the original AW RN inventory IRAS.
}
\details{
Returns the plot (ggplot).
}
