% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_input_parameters_combined_plots.r
\name{plot_irradiation_parameters}
\alias{plot_irradiation_parameters}
\title{Function to plot irradiation conditions: beam intensity,
irradiation time and waiting time. Returns the plot (patchwork wrap).}
\usage{
plot_irradiation_parameters(df)
}
\arguments{
\item{df}{Dataframe with beam intensity, irradiation and waiting time columns (can be scn or itm)}
}
\description{
Function to plot irradiation conditions: beam intensity,
irradiation time and waiting time. Returns the plot (patchwork wrap).
}
