% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_input_parameters_single_plots.R
\name{plot_machine}
\alias{plot_machine}
\title{Function to plot machine of origin barlot, in case it's needed.
Returns the plot (ggplot)}
\usage{
plot_machine(df)
}
\arguments{
\item{df}{dataframe with machine column (be it scn, itm or whatever works)}
}
\description{
Function to plot machine of origin barlot, in case it's needed.
Returns the plot (ggplot)
}
