% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_item_parameters.R
\name{plot_measures}
\alias{plot_measures}
\title{Function to plot measurable quantities: bgo counting rate and
dose rate.}
\usage{
plot_measures(df)
}
\arguments{
\item{df}{Dataframe with bgo and dose rate columns}
}
\description{
Returns the plot (patchwork wrap)
}
