% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_input_parameters_single_plots.R
\name{plot_waiting_time}
\alias{plot_waiting_time}
\title{Function to plot the waiting time distribution
Returns the plot (ggplot)}
\usage{
plot_waiting_time(df)
}
\arguments{
\item{df}{dataframe with waiting time column (be it scn, itm or whatever works)}
}
\description{
Function to plot the waiting time distribution
Returns the plot (ggplot)
}
