% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_theme.R
\docType{data}
\name{theme_no_y_axis}
\alias{theme_no_y_axis}
\title{Theme function that deletes the y axis of a plot
Use by adding +theme_no_y_axis to your ggplot object}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 4.
}
\usage{
theme_no_y_axis
}
\description{
Theme function that deletes the y axis of a plot
Use by adding +theme_no_y_axis to your ggplot object
}
\keyword{datasets}
