% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_input_parameters_combined_plots.r
\name{viz_all_input_parameters}
\alias{viz_all_input_parameters}
\title{Inspect all scenario parameters in a glance
Returns a named list of plots: "plot_physical_characteristics"
and "plot_irradiation_characteristics"}
\usage{
viz_all_input_parameters(df)
}
\arguments{
\item{df}{dataframe name}
}
\description{
Inspect all scenario parameters in a glance
Returns a named list of plots: "plot_physical_characteristics"
and "plot_irradiation_characteristics"
}
